requirements.txt
torch
torchvision
numpy
opencv-python
scikit-learn
pillow
tqdm
facenet-pytorch
matplotlib